backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0

python-dotenv>=1.0.1
requests>=2.31.0

# Mongo (Atlas)
motor==3.6.0
pymongo>=4.9.2
certifi>=2024.2.2

# Auth / Security
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
cryptography>=42.0.8
python-jose>=3.3.0
email-validator>=2.2.0

# Forms / uploads
python-multipart>=0.0.9

# Timezone data (useful on Linux)
tzdata>=2024.2

sib-api-v3-sdk>=7.0.0

# Storage
boto3==1.34.0
botocore==1.34.0

# Cloudinary
cloudinary>=1.36.0